backend/requirements.txt
Django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
google-auth==2.40.3
mysqlclient==2.2.8
oracledb==3.4.2
python-decouple==3.8
Pillow==11.2.1
inflection==0.5.1
requests==2.32.3
